// src/app/types/terrains.ts
export enum EPlanetTerrains {
  DESERT = "desert",
  ACID_POOLS = "acid pools",
  CLIFFS = "cliffs",
  SAVANNAHS = "savannahs",
  ROCK = "rock",
  MOUNTAINS = "mountains",
  SCRUBLANDS = "scrublands",
  GLACIERS = "glaciers",
  CANYONS = "canyons",
  TUNDRA = "tundra",
  ROCKY_CANYONS = "rocky canyons",
  FUNGUS_FORESTS = "fungus forests",
  JUNGLES = "jungles",
  SAVANNAS = "savannas",
  CITYSCAPE = "cityscape",
  GRASSY_HILLS = "grassy hills",
  SEAS = "seas",
  ROCKY_DESERTS = "rocky deserts",
  BARREN = "barren",
  AIRLESS_ASTEROID = "airless asteroid",
  FIELDS = "fields",
  CAVES = "caves",
  SWAMPS = "swamps",
  ASH = "ash",
  CITIES = "cities",
  MESAS = "mesas",
  MOUNTAIN_RANGES = "mountain ranges",
  URBAN = "urban",
  PLAINS = "plains",
  VERDANT = "verdant",
  MOUNTAIN = "mountain",
  GRASSLANDS = "grasslands",
  PLATEAUS = "plateaus",
  VINES = "vines",
  VOLCANOES = "volcanoes",
  ICE_CANYONS = "ice canyons",
  TOXIC_CLOUDSEA = "toxic cloudsea",
  DESERTS = "deserts",
  ICE_CAVES = "ice caves",
  ROCK_ARCHES = "rock arches",
  HILLS = "hills",
  ROCKY = "rocky",
  RIVERS = "rivers",
  ROCKY_ISLANDS = "rocky islands",
  JUNGLE = "jungle",
  GAS_GIANT = "gas giant",
  VALLEYS = "valleys",
  SWAMP = "swamp",
  ISLANDS = "islands",
  FORESTS = "forests",
  LAVA_RIVERS = "lava rivers",
  REEFS = "reefs",
  BOGS = "bogs",
  OCEAN = "ocean",
  LAKES = "lakes",
  UNKNOWN = "unknown",
  RAINFORESTS = "rainforests",
  OCEANS = "oceans",
  GRASS = "grass",
  SINKHOLES = "sinkholes",
}